requirements.txt
psycopg2-binary == 2.9.6
Django==4.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt
djoser
python-dotenv==1.0.0